Move stories & tests into drafts
