+++ src/drafts/InlineAutocomplete/InlineAutocomplete.test.tsx
@@ -1,11 +1,11 @@
 import React, {useLayoutEffect, useState} from 'react'
 import {fireEvent, render, within} from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
-import InlineAutocomplete, {ShowSuggestionsEvent, Suggestions, Trigger} from '../drafts/InlineAutocomplete'
-import FormControl from '../FormControl'
-import {ActionList} from '../ActionList'
-import Textarea from '../Textarea'
-import ThemeProvider from '../ThemeProvider'
+import InlineAutocomplete, {ShowSuggestionsEvent, Suggestions, Trigger} from '.'
+import FormControl from '../../FormControl'
+import {ActionList} from '../../ActionList'
+import Textarea from '../../Textarea'
+import ThemeProvider from '../../ThemeProvider'
 
 const label = 'Inline Autocomplete'
